Refactor OOB power management into dedicated modules
Pure structural refactor of the power-management feature; no behavior,
route-path, or visual changes.

- Director: move power logic out of director/mod.rs into power/ops.rs
  (PowerAction + power-kick helpers) and power_ops.rs (Director power
  method impls). crate::director::PowerAction and all Director method
  signatures preserved.
- HTTP: move the /ui/devices/{uuid}/power endpoints out of the
  oversized devices.rs into http/ui/power.rs, merged via power::routes.
- UI: extract the power badge, confirm dialog, and controls from
  BmcConfiguration.tsx into power-controls.tsx (538 -> 299 lines).
