+++ rack-director/src/director/mod.rs
@@ -3,12 +3,15 @@ use uuid::Uuid;
 use crate::database::Connection;
 use crate::director::store::generate_hostname_from_uuid;
 use crate::lifecycle::{DeviceLifecycle, LifecycleManager, LifecycleTransition, TransitionType};
-use crate::plans::actions::{Action, BootTarget};
+use crate::plans::actions::BootTarget;
 use crate::plans::{Plan, PlanStatus};
 use crate::{platforms, roles};
 
 pub(crate) mod power;
+mod power_ops;
 pub(crate) mod store;
+
+pub use power::PowerAction;
 
 pub use common::device_attributes::NetworkInterface;
 pub use store::Device;
@@ -39,119 +42,6 @@ impl Architecture {
 impl std::fmt::Display for Architecture {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         write!(f, "{}", self.as_str())
-    }
-}
-
-/// User-requested power action for the UI power controls.
-#[derive(Debug, Clone, Copy, serde::Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum PowerAction {
-    /// Power the device on.
-    On,
-    /// Power the device off (hard/immediate power-off, not a graceful OS shutdown).
-    ///
-    /// A hard off is used because hosts are frequently running the rack-agent in
-    /// an initramfs that does not handle ACPI soft-off, so a graceful shutdown can
-    /// silently hang. This also matches the UI confirm dialog's promise that the
-    /// device will lose power immediately.
-    Off,
-    /// Power-cycle the device (off then on).
-    Cycle,
-}
-
-/// Returns `true` if `action` requires the device to be booted into the rack-agent
-/// before it can be executed.
-///
-/// Actions that need the agent running on the device return `true` and will trigger an
-/// OOB power kick when the device is not already in daemon mode. `RebootDevice` returns
-/// `false` because its `start()` hook already handles the reboot via `Director::reboot`.
-pub(crate) fn action_requires_boot(action: &Action) -> bool {
-    matches!(
-        action,
-        Action::DiscoverHardware
-            | Action::ConfigureBmc
-            | Action::PartitionDisks
-            | Action::InstallOs
-            | Action::Console
-    )
-}
-
-/// Apply a power kick using the already-resolved driver.
-///
-/// Queries the current power state and issues the appropriate command:
-/// - `Off`        → `power_on`
-/// - `On`         → `power_cycle` (reprovision path — device likely running an OS)
-/// - `Unknown`    → `power_cycle`, with a `power_on` fallback if the cycle fails
-/// - state query error → treated as `Unknown` (log warn + `power_cycle` with fallback)
-///
-/// The `Unknown` fallback exists because Redfish `ForceRestart` (and IPMI
-/// `chassis power cycle`) typically error when the host is actually OFF. In that
-/// case a plain `power_on` is the correct recovery, so we attempt it rather than
-/// giving up after a single failed cycle. The confirmed-`On` case does NOT fall
-/// back to `power_on`: a device that is genuinely on should not be powered on
-/// again if its cycle command fails.
-///
-/// Every power operation is best-effort: an error is logged but never propagated.
-/// Returns `Ok(())` in all cases.
-pub(crate) async fn apply_power_kick(driver: &dyn power::PowerDriver) -> anyhow::Result<()> {
-    let state = match driver.power_state().await {
-        Ok(s) => s,
-        Err(e) => {
-            log::warn!(
-                "Could not query power state via {} driver: {} — treating as Unknown",
-                driver.kind(),
-                e
-            );
-            power::PowerState::Unknown
-        }
-    };
-
-    match state {
-        power::PowerState::Off => {
-            log::info!(
-                "Device is off — issuing power_on via {} driver",
-                driver.kind()
-            );
-            log_if_err(driver.power_on().await, driver.kind());
-        }
-        power::PowerState::On => {
-            log::info!(
-                "Device is On — issuing power_cycle via {} driver",
-                driver.kind()
-            );
-            log_if_err(driver.power_cycle().await, driver.kind());
-        }
-        power::PowerState::Unknown => cycle_with_power_on_fallback(driver).await,
-    }
-
-    Ok(())
-}
-
-/// Issue a `power_cycle` for a device in an `Unknown` state, falling back to
-/// `power_on` if the cycle fails.
-///
-/// A failed cycle on an Unknown host most often means the host is actually OFF
-/// (Redfish `ForceRestart` errors in that state), so `power_on` is the correct
-/// best-effort recovery. Both operations log on failure but never propagate.
-async fn cycle_with_power_on_fallback(driver: &dyn power::PowerDriver) {
-    log::info!(
-        "Device is Unknown — issuing power_cycle via {} driver",
-        driver.kind()
-    );
-    if let Err(e) = driver.power_cycle().await {
-        log::warn!(
-            "power_cycle failed via {} driver: {} — falling back to power_on",
-            driver.kind(),
-            e
-        );
-        log_if_err(driver.power_on().await, driver.kind());
-    }
-}
-
-/// Log a best-effort power command failure without propagating it.
-fn log_if_err(result: anyhow::Result<()>, kind: &str) {
-    if let Err(e) = result {
-        log::warn!("Power kick command failed via {} driver: {}", kind, e);
     }
 }
 
@@ -1008,157 +898,6 @@ impl<'a> Director<'a> {
         store::find_device_by_bmc_mac(self.conn, mac).await
     }
 
-    /// Ensure the device is powered and will boot for the given plan action.
-    ///
-    /// Called immediately before `action.start()` in `start_lifecycle_transition`.
-    /// Short-circuits without any BMC call when:
-    /// 1. The action does not require the device to boot (e.g. `RebootDevice`).
-    /// 2. The agent is already polling in daemon mode (heartbeat within `DAEMON_HEARTBEAT_WINDOW`).
-    /// 3. The device has no BMC configured.
-    ///
-    /// All power operations are **best-effort**: errors are logged but never propagated.
-    /// Always returns `Ok(())`.
-    pub async fn ensure_powered_for_plan(
-        &self,
-        device: &store::Device,
-        action: &Action,
-    ) -> anyhow::Result<()> {
-        if !action_requires_boot(action) {
-            return Ok(());
-        }
-
-        if power::is_in_daemon_mode(
-            device.last_polled_at.as_deref(),
-            power::DAEMON_HEARTBEAT_WINDOW,
-        ) {
-            log::info!(
-                "Device {} is in daemon mode (last_polled_at={}), skipping power kick",
-                device.uuid,
-                device.last_polled_at.as_deref().unwrap_or("none")
-            );
-            return Ok(());
-        }
-
-        let driver = match self.power_driver_for(&device.uuid).await? {
-            Some(d) => d,
-            None => {
-                log::info!(
-                    "No BMC driver available for device {}, skipping power kick",
-                    device.uuid
-                );
-                return Ok(());
-            }
-        };
-
-        if let Err(e) = apply_power_kick(driver.as_ref()).await {
-            log::warn!(
-                "apply_power_kick returned error for device {}: {}",
-                device.uuid,
-                e
-            );
-        }
-
-        Ok(())
-    }
-
-    /// Issue a power reset to the device's BMC using the best available driver.
-    ///
-    /// Probes the BMC for Redfish support and falls back to IPMI.  This is a
-    /// best-effort operation: if the BMC IP or credentials are missing, or if
-    /// the power command fails, the error is logged but not propagated.  This
-    /// allows lifecycle transitions to continue even when OOB power is unavailable.
-    pub async fn reboot(&self, uuid: &Uuid) -> anyhow::Result<()> {
-        match self.power_driver_for(uuid).await? {
-            Some(driver) => {
-                if let Err(e) = driver.power_reset().await {
-                    log::warn!(
-                        "Power reset failed for device {} via {} driver: {}",
-                        uuid,
-                        driver.kind(),
-                        e
-                    );
-                }
-            }
-            None => {
-                log::info!(
-                    "No BMC configured for device {}, skipping power reset",
-                    uuid
-                );
-            }
-        }
-        Ok(())
-    }
-
-    /// Resolve the best available power driver for a device.
-    ///
-    /// Looks up the device's BMC IP and credentials, then calls
-    /// [`power::resolve_power_driver`] to probe for Redfish support and fall
-    /// back to IPMI if unavailable.
-    ///
-    /// Returns `Ok(None)` when the device has no BMC IP or no credentials
-    /// stored — callers should log and skip the power operation in that case.
-    ///
-    /// This is the foundation for future power endpoints and the kick logic.
-    /// Callers that need power operations should call this method and act on
-    /// the returned driver.
-    pub async fn power_driver_for(
-        &self,
-        uuid: &Uuid,
-    ) -> anyhow::Result<Option<Box<dyn power::PowerDriver>>> {
-        let ip = match self.get_bmc_ip(uuid).await? {
-            Some(ip) => ip,
-            None => return Ok(None),
-        };
-        let (username, password) = match self.get_bmc_credentials(uuid).await? {
-            Some(creds) => creds,
-            None => return Ok(None),
-        };
-        Ok(Some(
-            power::resolve_power_driver(&ip, &username, &password, self.power_config).await,
-        ))
-    }
-
-    /// Query the current power state and driver kind for a device.
-    ///
-    /// Never errors — a missing BMC, unreachable BMC, or driver failure all
-    /// degrade to `(PowerState::Unknown, None)` so that the UI power badge
-    /// can always render without blocking the page.
-    pub async fn power_status(&self, uuid: &Uuid) -> (power::PowerState, Option<String>) {
-        match self.power_driver_for(uuid).await {
-            Ok(Some(d)) => {
-                let state = d.power_state().await.unwrap_or(power::PowerState::Unknown);
-                (state, Some(d.kind().to_string()))
-            }
-            _ => (power::PowerState::Unknown, None),
-        }
-    }
-
-    /// Execute a UI-requested power action on a device.
-    ///
-    /// `PowerAction::Off` issues a **hard** (immediate) power-off — not a
-    /// graceful OS shutdown — because hosts often run the rack-agent in an
-    /// initramfs that cannot honor ACPI soft-off, and this matches the UI
-    /// confirm dialog's promise of immediate power loss.
-    ///
-    /// Returns:
-    /// - `Ok(true)`  – action was issued successfully.
-    /// - `Ok(false)` – device has no BMC configured (caller should return 404).
-    /// - `Err(_)`    – BMC driver reported a failure (caller should return 502).
-    pub async fn power_action(&self, uuid: &Uuid, action: PowerAction) -> anyhow::Result<bool> {
-        let driver = match self.power_driver_for(uuid).await? {
-            Some(d) => d,
-            None => return Ok(false),
-        };
-
-        match action {
-            PowerAction::On => driver.power_on().await?,
-            PowerAction::Off => driver.power_off(false).await?,
-            PowerAction::Cycle => driver.power_cycle().await?,
-        }
-
-        Ok(true)
-    }
-
     /// Cancel the active lifecycle transition for a device.
     ///
     /// Uses a CAS-style conditional UPDATE on the plan so that a concurrent
@@ -1205,26 +944,6 @@ impl<'a> Director<'a> {
         .await?;
 
         Ok(())
-    }
-
-    /// Get BMC IP address from device attributes
-    async fn get_bmc_ip(&self, uuid: &Uuid) -> anyhow::Result<Option<String>> {
-        let device = store::get_device(self.conn, uuid).await?;
-        Ok(device.attributes.bmc.and_then(|bmc| bmc.ip_address))
-    }
-
-    /// Get BMC credentials from device attributes
-    async fn get_bmc_credentials(&self, uuid: &Uuid) -> anyhow::Result<Option<(String, String)>> {
-        let device = store::get_device(self.conn, uuid).await?;
-        let bmc_config = device.attributes.bmc_config;
-
-        match bmc_config {
-            Some(config) => match (config.username, config.password) {
-                (Some(u), Some(p)) => Ok(Some((u, p))),
-                _ => Ok(None),
-            },
-            None => Ok(None),
-        }
     }
 }
 
@@ -1462,94 +1181,6 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_reboot_with_bmc_info() {
-        let conn = setup_test_db(test_connection_factory!()).await;
-        let director = Director::new(&conn);
-        let test_uuid = Uuid::parse_str("550e8400-e29b-41d4-a716-446655440020").unwrap();
-
-        // Register device
-        director
-            .register_device(&test_uuid, Architecture::X86_64)
-            .await
-            .unwrap();
-
-        // Set BMC info and credentials
-        let mut attributes = serde_json::Map::new();
-        attributes.insert(
-            "bmc".to_string(),
-            serde_json::json!({
-                "mac_address": "aa:bb:cc:dd:ee:ff",
-                "ip_address": "10.0.0.100"
-            }),
-        );
-        attributes.insert(
-            "bmc_config".to_string(),
-            serde_json::json!({
-                "ip_address_source": "static",
-                "username": "RACKDIRECTOR",
-                "password": "test_password"
-            }),
-        );
-        director
-            .update_attributes(&test_uuid, attributes)
-            .await
-            .unwrap();
-
-        // Call reboot - should not fail even if ipmitool is not installed
-        // (it's best-effort)
-        let result = director.reboot(&test_uuid).await;
-        assert!(result.is_ok());
-    }
-
-    #[tokio::test]
-    async fn test_reboot_without_bmc_ip() {
-        let conn = setup_test_db(test_connection_factory!()).await;
-        let director = Director::new(&conn);
-        let test_uuid = Uuid::parse_str("550e8400-e29b-41d4-a716-446655440021").unwrap();
-
-        // Register device without BMC info
-        director
-            .register_device(&test_uuid, Architecture::X86_64)
-            .await
-            .unwrap();
-
-        // Call reboot - should succeed (gracefully skip IPMI)
-        let result = director.reboot(&test_uuid).await;
-        assert!(result.is_ok());
-    }
-
-    #[tokio::test]
-    async fn test_reboot_without_bmc_credentials() {
-        let conn = setup_test_db(test_connection_factory!()).await;
-        let director = Director::new(&conn);
-        let test_uuid = Uuid::parse_str("550e8400-e29b-41d4-a716-446655440022").unwrap();
-
-        // Register device
-        director
-            .register_device(&test_uuid, Architecture::X86_64)
-            .await
-            .unwrap();
-
-        // Set BMC IP but no credentials
-        let mut attributes = serde_json::Map::new();
-        attributes.insert(
-            "bmc".to_string(),
-            serde_json::json!({
-                "mac_address": "aa:bb:cc:dd:ee:ff",
-                "ip_address": "10.0.0.100"
-            }),
-        );
-        director
-            .update_attributes(&test_uuid, attributes)
-            .await
-            .unwrap();
-
-        // Call reboot - should succeed (gracefully skip IPMI)
-        let result = director.reboot(&test_uuid).await;
-        assert!(result.is_ok());
-    }
-
-    #[tokio::test]
     async fn test_on_boot_advances_reboot_action() {
         let conn = setup_test_db(test_connection_factory!()).await;
         let director = Director::new(&conn);
@@ -1658,106 +1289,6 @@ mod tests {
         // Call on_boot - should succeed without error
         let result = director.on_boot(&test_uuid).await;
         assert!(result.is_ok());
-    }
-
-    #[tokio::test]
-    async fn test_get_bmc_ip_with_valid_bmc() {
-        let conn = setup_test_db(test_connection_factory!()).await;
-        let director = Director::new(&conn);
-        let test_uuid = Uuid::parse_str("550e8400-e29b-41d4-a716-446655440026").unwrap();
-
-        // Register device
-        director
-            .register_device(&test_uuid, Architecture::X86_64)
-            .await
-            .unwrap();
-
-        // Set BMC IP
-        let mut attributes = serde_json::Map::new();
-        attributes.insert(
-            "bmc".to_string(),
-            serde_json::json!({
-                "mac_address": "aa:bb:cc:dd:ee:ff",
-                "ip_address": "10.0.0.100"
-            }),
-        );
-        director
-            .update_attributes(&test_uuid, attributes)
-            .await
-            .unwrap();
-
-        // Get BMC IP
-        let bmc_ip = director.get_bmc_ip(&test_uuid).await.unwrap();
-        assert_eq!(bmc_ip, Some("10.0.0.100".to_string()));
-    }
-
-    #[tokio::test]
-    async fn test_get_bmc_ip_without_bmc() {
-        let conn = setup_test_db(test_connection_factory!()).await;
-        let director = Director::new(&conn);
-        let test_uuid = Uuid::parse_str("550e8400-e29b-41d4-a716-446655440027").unwrap();
-
-        // Register device without BMC
-        director
-            .register_device(&test_uuid, Architecture::X86_64)
-            .await
-            .unwrap();
-
-        // Get BMC IP
-        let bmc_ip = director.get_bmc_ip(&test_uuid).await.unwrap();
-        assert_eq!(bmc_ip, None);
-    }
-
-    #[tokio::test]
-    async fn test_get_bmc_credentials_with_valid_config() {
-        let conn = setup_test_db(test_connection_factory!()).await;
-        let director = Director::new(&conn);
-        let test_uuid = Uuid::parse_str("550e8400-e29b-41d4-a716-446655440028").unwrap();
-
-        // Register device
-        director
-            .register_device(&test_uuid, Architecture::X86_64)
-            .await
-            .unwrap();
-
-        // Set BMC credentials
-        let mut attributes = serde_json::Map::new();
-        attributes.insert(
-            "bmc_config".to_string(),
-            serde_json::json!({
-                "ip_address_source": "dhcp",
-                "username": "RACKDIRECTOR",
-                "password": "test_password"
-            }),
-        );
-        director
-            .update_attributes(&test_uuid, attributes)
-            .await
-            .unwrap();
-
-        // Get BMC credentials
-        let creds = director.get_bmc_credentials(&test_uuid).await.unwrap();
-        assert_eq!(
-            creds,
-            Some(("RACKDIRECTOR".to_string(), "test_password".to_string()))
-        );
-    }
-
-    #[tokio::test]
-    async fn test_get_bmc_credentials_without_credentials() {
-        let conn = setup_test_db(test_connection_factory!()).await;
-        let director = Director::new(&conn);
-        let test_uuid = Uuid::parse_str("550e8400-e29b-41d4-a716-446655440029").unwrap();
-
-        // Register device - will have default BMC config with no credentials
-        director
-            .register_device(&test_uuid, Architecture::X86_64)
-            .await
-            .unwrap();
-
-        // Get BMC credentials - should be None even though config exists
-        let creds = director.get_bmc_credentials(&test_uuid).await.unwrap();
-        assert_eq!(creds, None);
     }
 
     #[tokio::test]
@@ -2939,274 +2470,6 @@ mod tests {
                 .to_string()
                 .contains("No active transition to cancel")
         );
-    }
-
-    // ========== action_requires_boot table tests ==========
-
-    #[test]
-    fn test_action_requires_boot_discover_hardware() {
-        assert!(action_requires_boot(&Action::DiscoverHardware));
-    }
-
-    #[test]
-    fn test_action_requires_boot_configure_bmc() {
-        assert!(action_requires_boot(&Action::ConfigureBmc));
-    }
-
-    #[test]
-    fn test_action_requires_boot_partition_disks() {
-        assert!(action_requires_boot(&Action::PartitionDisks));
-    }
-
-    #[test]
-    fn test_action_requires_boot_install_os() {
-        assert!(action_requires_boot(&Action::InstallOs));
-    }
-
-    #[test]
-    fn test_action_requires_boot_console() {
-        assert!(action_requires_boot(&Action::Console));
-    }
-
-    #[test]
-    fn test_action_requires_boot_reboot_device_is_false() {
-        assert!(!action_requires_boot(&Action::RebootDevice));
-    }
-
-    // ========== apply_power_kick tests with FakeDriver ==========
-
-    use crate::director::power::{PowerDriver, PowerState};
-    use std::sync::Mutex;
-
-    /// A fake `PowerDriver` that records every command issued (in order) and
-    /// returns a configurable state from `power_state`.
-    struct FakeDriver {
-        state: PowerState,
-        /// Tracks every power command issued, in order: "on", "off", "cycle", "reset".
-        commands: Mutex<Vec<String>>,
-        /// If true, `power_state()` returns an error instead.
-        state_error: bool,
-        /// If true, `power_cycle()` returns an error instead of succeeding.
-        cycle_fails: bool,
-    }
-
-    impl FakeDriver {
-        fn new(state: PowerState) -> Self {
-            Self {
-                state,
-                commands: Mutex::new(Vec::new()),
-                state_error: false,
-                cycle_fails: false,
-            }
-        }
-
-        fn with_state_error() -> Self {
-            Self {
-                state: PowerState::Unknown,
-                commands: Mutex::new(Vec::new()),
-                state_error: true,
-                cycle_fails: false,
-            }
-        }
-
-        /// Build a driver in `state` whose `power_cycle()` always fails.
-        fn with_cycle_failure(state: PowerState) -> Self {
-            Self {
-                state,
-                commands: Mutex::new(Vec::new()),
-                state_error: false,
-                cycle_fails: true,
-            }
-        }
-
-        fn record(&self, command: &str) {
-            self.commands.lock().unwrap().push(command.to_string());
-        }
-
-        /// All commands issued so far, in order.
-        fn commands(&self) -> Vec<String> {
-            self.commands.lock().unwrap().clone()
-        }
-
-        /// The most recent command issued, if any.
-        fn last_command(&self) -> Option<String> {
-            self.commands.lock().unwrap().last().cloned()
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl PowerDriver for FakeDriver {
-        async fn power_state(&self) -> anyhow::Result<PowerState> {
-            if self.state_error {
-                Err(anyhow::anyhow!("simulated power_state error"))
-            } else {
-                Ok(self.state)
-            }
-        }
-
-        async fn power_on(&self) -> anyhow::Result<()> {
-            self.record("on");
-            Ok(())
-        }
-
-        async fn power_off(&self, _graceful: bool) -> anyhow::Result<()> {
-            self.record("off");
-            Ok(())
-        }
-
-        async fn power_cycle(&self) -> anyhow::Result<()> {
-            self.record("cycle");
-            if self.cycle_fails {
-                Err(anyhow::anyhow!("simulated power_cycle failure"))
-            } else {
-                Ok(())
-            }
-        }
-
-        async fn power_reset(&self) -> anyhow::Result<()> {
-            self.record("reset");
-            Ok(())
-        }
-
-        fn kind(&self) -> &'static str {
-            "fake"
-        }
-    }
-
-    #[tokio::test]
-    async fn test_apply_power_kick_when_off_calls_power_on() {
-        let driver = FakeDriver::new(PowerState::Off);
-        apply_power_kick(&driver).await.unwrap();
-        assert_eq!(driver.last_command(), Some("on".to_string()));
-    }
-
-    #[tokio::test]
-    async fn test_apply_power_kick_when_on_calls_power_cycle() {
-        let driver = FakeDriver::new(PowerState::On);
-        apply_power_kick(&driver).await.unwrap();
-        assert_eq!(driver.last_command(), Some("cycle".to_string()));
-    }
-
-    #[tokio::test]
-    async fn test_apply_power_kick_when_unknown_calls_power_cycle() {
-        let driver = FakeDriver::new(PowerState::Unknown);
-        apply_power_kick(&driver).await.unwrap();
-        assert_eq!(driver.last_command(), Some("cycle".to_string()));
-    }
-
-    #[tokio::test]
-    async fn test_apply_power_kick_when_state_error_calls_power_cycle() {
-        let driver = FakeDriver::with_state_error();
-        apply_power_kick(&driver).await.unwrap();
-        assert_eq!(driver.last_command(), Some("cycle".to_string()));
-    }
-
-    #[tokio::test]
-    async fn test_apply_power_kick_unknown_cycle_fails_falls_back_to_power_on() {
-        // An Unknown host whose power_cycle fails is most likely actually OFF
-        // (Redfish ForceRestart errors when off) — we should fall back to power_on.
-        let driver = FakeDriver::with_cycle_failure(PowerState::Unknown);
-        apply_power_kick(&driver).await.unwrap();
-        assert_eq!(
-            driver.commands(),
-            vec!["cycle".to_string(), "on".to_string()],
-            "Unknown + cycle failure should attempt power_cycle then fall back to power_on"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_apply_power_kick_on_cycle_fails_does_not_fall_back() {
-        // A confirmed-On host whose power_cycle fails must NOT be powered on again.
-        let driver = FakeDriver::with_cycle_failure(PowerState::On);
-        apply_power_kick(&driver).await.unwrap();
-        assert_eq!(
-            driver.commands(),
-            vec!["cycle".to_string()],
-            "On + cycle failure should NOT fall back to power_on"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_apply_power_kick_returns_ok_even_if_power_op_fails() {
-        // A driver where every power command fails
-        struct FailDriver;
-        #[async_trait::async_trait]
-        impl PowerDriver for FailDriver {
-            async fn power_state(&self) -> anyhow::Result<PowerState> {
-                Ok(PowerState::On)
-            }
-            async fn power_on(&self) -> anyhow::Result<()> {
-                Err(anyhow::anyhow!("fail"))
-            }
-            async fn power_off(&self, _: bool) -> anyhow::Result<()> {
-                Err(anyhow::anyhow!("fail"))
-            }
-            async fn power_cycle(&self) -> anyhow::Result<()> {
-                Err(anyhow::anyhow!("fail"))
-            }
-            async fn power_reset(&self) -> anyhow::Result<()> {
-                Err(anyhow::anyhow!("fail"))
-            }
-            fn kind(&self) -> &'static str {
-                "fail"
-            }
-        }
-
-        // Even when the power command fails, apply_power_kick returns Ok(())
-        let result = apply_power_kick(&FailDriver).await;
-        assert!(result.is_ok());
-    }
-
-    // ========== ensure_powered_for_plan daemon-mode skip test ==========
-
-    #[tokio::test]
-    async fn test_ensure_powered_for_plan_skips_when_in_daemon_mode() {
-        let conn = setup_test_db(test_connection_factory!()).await;
-        let director = Director::new(&conn);
-        let test_uuid = Uuid::parse_str("550e8400-e29b-41d4-a716-446655440070").unwrap();
-
-        director
-            .register_device(&test_uuid, Architecture::X86_64)
-            .await
-            .unwrap();
-
-        // Manually set last_polled_at to now (within the 15s window) so
-        // is_in_daemon_mode returns true and the kick is skipped.
-        conn.execute(
-            "UPDATE devices SET last_polled_at = CURRENT_TIMESTAMP WHERE uuid = ?1",
-            (test_uuid,),
-        )
-        .await
-        .unwrap();
-
-        let device = store::get_device(&conn, &test_uuid).await.unwrap();
-
-        // DiscoverHardware requires boot, but daemon mode is active → should skip
-        let result = director
-            .ensure_powered_for_plan(&device, &Action::DiscoverHardware)
-            .await;
-        assert!(result.is_ok());
-        // No BMC configured either, but we never reach that check due to daemon mode
-    }
-
-    #[tokio::test]
-    async fn test_ensure_powered_for_plan_skips_reboot_device() {
-        let conn = setup_test_db(test_connection_factory!()).await;
-        let director = Director::new(&conn);
-        let test_uuid = Uuid::parse_str("550e8400-e29b-41d4-a716-446655440071").unwrap();
-
-        director
-            .register_device(&test_uuid, Architecture::X86_64)
-            .await
-            .unwrap();
-
-        let device = store::get_device(&conn, &test_uuid).await.unwrap();
-
-        // RebootDevice does not require boot → ensure_powered_for_plan returns immediately
-        let result = director
-            .ensure_powered_for_plan(&device, &Action::RebootDevice)
-            .await;
-        assert!(result.is_ok());
     }
 
     #[tokio::test]
